# Support differential/velocity data in frame concatenate() function 