--- a/astropy/coordinates/funcs.py
+++ b/astropy/coordinates/funcs.py
@@ -8,6 +8,7 @@ framework, but it is useful for some users who are used to more functional
 interfaces.
 """
 
+from collections.abc import Sequence
 
 import numpy as np
 
@@ -168,32 +169,6 @@ def get_sun(time):
     return SkyCoord(cartrep, frame=GCRS(obstime=time))
 
 
-def concatenate(coords):
-    """
-    Combine multiple coordinate objects into a single
-    `~astropy.coordinates.SkyCoord`.
-
-    "Coordinate objects" here mean frame objects with data,
-    `~astropy.coordinates.SkyCoord`, or representation objects.  Currently,
-    they must all be in the same frame, but in a future version this may be
-    relaxed to allow inhomogenous sequences of objects.
-
-    Parameters
-    ----------
-    coords : sequence of coordinate objects
-        The objects to concatenate
-
-    Returns
-    -------
-    cskycoord : SkyCoord
-        A single sky coordinate with its data set to the concatenation of all
-        the elements in ``coords``
-    """
-    if getattr(coords, 'isscalar', False) or not isiterable(coords):
-        raise TypeError('The argument to concatenate must be iterable')
-    return SkyCoord(coords)
-
-
 # global dictionary that caches repeatedly-needed info for get_constellation
 _constellation_data = {}
 
@@ -279,3 +254,104 @@ def get_constellation(coord, short_name=False, constellation_list='iau'):
         return names[0]
     else:
         return names
+
+
+def _concatenate_components(reps_difs, names):
+    """ Helper function for the concatenate function below. Gets and
+    concatenates all of the individual components for an iterable of
+    representations or differentials.
+    """
+    values = []
+    for name in names:
+        data_vals = []
+        for x in reps_difs:
+            data_val = getattr(x, name)
+            data_vals.append(data_val.reshape(1, ) if x.isscalar else data_val)
+        concat_vals = np.concatenate(data_vals)
+
+        # Hack because np.concatenate doesn't fully work with Quantity
+        if isinstance(concat_vals, u.Quantity):
+            concat_vals._unit = data_val.unit
+
+        values.append(concat_vals)
+
+    return values
+
+def concatenate_representations(reps):
+    """
+    Combine multiple representation objects into a single instance by
+    concatenating the data in each component.
+
+    Currently, all of the input representations have to be the same type. This
+    properly handles differential or velocity data, but all input objects must
+    have the same differential object type as well.
+
+    Parameters
+    ----------
+    reps : sequence of representation objects
+        The objects to concatenate
+
+    Returns
+    -------
+    rep : `~astropy.coordinates.BaseRepresentation` subclass
+        A single representation object with its data set to the concatenation of
+        all the elements of the input sequence of representations.
+    """
+    if not isinstance(reps, (Sequence, np.ndarray)):
+        raise TypeError('Input must be a list or iterable of representation '
+                        'objects.')
+
+    # First, validate that the represenations are the same, and
+    # concatenate all of the positional data:
+    rep_type = type(reps[0])
+    if any(type(r) != rep_type for r in reps):
+        raise ValueError('Input representations must all have the same type.')
+
+    # Construct the new representation with the concatenated data from the
+    # representations passed in
+    values = _concatenate_components(reps,
+                                     rep_type.attr_classes.keys())
+    new_rep = rep_type(*values)
+
+    has_diff = 's' in reps[0].differentials
+    if has_diff:
+        dif_type = type(reps[0].differentials['s'])
+
+        if any('s' not in r.differentials or
+                type(r.differentials['s']) != dif_type
+               for r in reps):
+            raise ValueError('Input representations must all have the same '
+                             'differential type.')
+
+        values = _concatenate_components([r.differentials['s'] for r in reps],
+                                         dif_type.attr_classes.keys())
+        new_dif = dif_type(*values)
+        new_rep = new_rep.with_differentials({'s': new_dif})
+
+    return new_rep
+
+
+def concatenate(coords):
+    """
+    Combine multiple coordinate objects into a single
+    `~astropy.coordinates.SkyCoord`.
+
+    "Coordinate objects" here mean frame objects with data,
+    `~astropy.coordinates.SkyCoord`, or representation objects.  Currently,
+    they must all be in the same frame, but in a future version this may be
+    relaxed to allow inhomogenous sequences of objects.
+
+    Parameters
+    ----------
+    coords : sequence of coordinate objects
+        The objects to concatenate
+
+    Returns
+    -------
+    cskycoord : SkyCoord
+        A single sky coordinate with its data set to the concatenation of all
+        the elements in ``coords``
+    """
+    if getattr(coords, 'isscalar', False) or not isiterable(coords):
+        raise TypeError('The argument to concatenate must be iterable')
+    return SkyCoord(coords)