This PR does two things:

1. Fixes #7506 - the `concatenate()` function now properly handles differential data in frames.
2. Adds a new function to concatenate raw `Representation` instances (with or without attached differentials), and offloads most of the work here.

Previously, all of the work in `concatenate()` happened in the `SkyCoord` initializer.