@adrn 2018-10-19
> does this have better, worse, or same performance vs current master?

Sounds like I should add a benchmark ;)