@adrn 2018-10-19
OK, I think I addressed the comments you left in here! 

On performance, this PR is ~5 times faster:

```python
c1 = coord.ICRS(ra=1*u.deg, dec=2*u.deg)
c2 = coord.ICRS(ra=np.random.random(10000)*u.deg, dec=np.random.random(10000)*u.deg)
```

This PR:
```python
%timeit coord.concatenate((c1, c1))
1.9 ms ± 408 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
%timeit coord.concatenate((c2, c2))
2.32 ms ± 388 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```

master:
```python
%timeit coord.concatenate((c1, c1))
8.97 ms ± 340 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
%timeit coord.concatenate((c2, c2))
10.1 ms ± 910 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
```